+++ client/src/schema/xataObject.ts
@@ -1,0 +1,11 @@
+import { Selectable } from './query';
+
+export interface XataObject {
+  id: string;
+  xata: {
+    version: number;
+  };
+  read(): Promise<this>;
+  update(data: Selectable<this>): Promise<this>;
+  delete(): Promise<void>;
+}
